viewsavedINV

+++ back/DTOs/InventoryCheckItemDTO.cs
@@ -1,5 +1,5 @@
 public class InventoryCheckItemDto
 {
     public int Id { get; set; }
-    public int RecordedAmount { get; set; }
+    public float RecordedAmount { get; set; }
 }
