+++ back/Models/InventoryCheckItem.cs
@@ -2,7 +2,7 @@ public class InventoryCheckItem
 {
     public int ItemId { get; set; }
     public int InventoryCheckId { get; set; }
-    public int RecordedAmount { get; set; }
+    public float RecordedAmount { get; set; }
 
     // Właściwości nawigacyjne
     public Item? Item { get; set; }
